# Ensure `Matrix` * `(Sym)Tridiagonal` is dense