This PR ensures that `Matrix` * `(Sym)Tridiagonal` outputs a dense matrix. Before, it was outputting a `SparseMatrixCSC`, despite the resulting matrix being dense.

This PR currently only implements this optimization for the `Matrix` type, so as to avoid method ambiguities, but it is possible that there are other type combinations that suffer from the same problem.

------------------------------------------------------------------------------------
A basic benchmark, showing a ~150x speedup:

Before:

```julia
julia> t = Tridiagonal(rand(1000, 1000)); s = SymTridiagonal(rand(1000), rand(999)); y = rand(1000, 1000);

julia> @btime $y * $t;
  256.924 ms (41 allocations: 18.01 MiB)

julia> @btime $y * $s;
  260.044 ms (41 allocations: 18.01 MiB)
```

This PR:

```julia
julia> t = Tridiagonal(rand(1000, 1000)); s = SymTridiagonal(rand(1000), rand(999)); y = rand(1000, 1000);

julia> @btime $y * $t;
  1.477 ms (2 allocations: 7.63 MiB)

julia> @btime $y * $s;
  1.469 ms (2 allocations: 7.63 MiB)
```