@dkarrasch 2021-12-13
Could it be that this is already handled by #43127?