@dkarrasch 2021-12-13
> Could it be that this is already handled by #43127?

Yes, it is, I checked. Good to know that there are also performance benefits from splitting LinearAlgebra and SparseArrays. Maybe we keep this PR for the added tests until #43127 is merged?