@stevengj 2021-12-17
Is there a similar method for `tridiagonal * matrix` that also needs to be fixed?