@mcognetta 2021-12-17
@stevengj Not for Tridiagonal * Dense, but perhaps for others. I think we are getting pretty close to catching all the defined structured matrix multiplication optimizations (though it always seems like one slips through the cracks).