@mcognetta 2021-12-26
Opened #43549 to resolve this.